refactor: logic and testing cleanup
